chore: split oversized backend files under 500-line hard cap
- backend/tests/test_api.py 509 → 400: extract TestRoleManagementAPI
  into backend/tests/test_role_management.py (12 tests, with their
  manage_users_* fixtures).
- backend/community/_docs.py 501 → 313: extract document CRUD + reorder
  endpoints into backend/community/_docs_documents.py. Shared schemas
  and _doc_to_out/_has_manage_docs helpers stay in _docs.py.
- backend/community/api.py: register the new docs_documents_router.

Pure reorg — no behavior change. All 699 backend tests pass.
